moved tests into a test folder
