+++ src/components/WriterOnboardingForms/tests/WriterEducationForm.test.js
@@ -2,9 +2,9 @@ import React, { useState as useStateMock } from "react";
 import { axe } from "jest-axe";
 import { createStore } from "redux";
 import { render as rtlRender, screen } from "@testing-library/react";
-import WriterEducationForm from "./WriterEducationForm.js";
-import { initialState as initialReducerState } from "../../store/reducers/collegesReducer";
-import reducer from "../../store/reducers/collegesReducer";
+import WriterEducationForm from "../WriterEducationForm.js";
+import { initialState as initialReducerState } from "../../../store/reducers/collegesReducer";
+import reducer from "../../../store/reducers/collegesReducer";
 
 import { Provider } from "react-redux";
